Another update of the CMakeLists.cmake and of the organisation of the files.
Now during the install all the headers are exported.
The install should fail if lp_solve has not been install on the system previously: if /usr/local/include/lp_solve/lp_lib.h does not exist, the comilation will fail when make will be looking for it while compiling the library StabilityPolytope.

I could not find a fix that work both for the compilation of stabiliplus and then once it is installed allows other project to compile...