[Medium] Fixed issue when mass is 0
The equilibrium region is mass independant, however in the way the equalities and inequalities were implemented, the mass appeared twice. Because of this when the given mass of the contact set was 0, it was creating issues. By remove the mass from the computation (set to 1 as it only appeared in multiplication) the issue is solved.